# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /onboarding/company:
    post:
      operationId: onboardCompany
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompanyOnboardingInput" }
      responses:
        "201": { description: Company created, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "400": { description: Invalid company name }
        "401": { description: Unauthorized }
        "409": { description: Already a member }
...
  /works:
    post:
      operationId: createWork
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkInput" } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Work" } } } }, "403": { description: Forbidden } }
  /expenses:
    post:
      operationId: createExpense
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ExpenseInput" } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Expense" } } } }, "403": { description: Forbidden } }
  /incidents:
    post:
      operationId: createIncident
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/IncidentInput" } } } }
      responses: { "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } } }, "403": { description: Forbidden } }
  /incidents/{id}/status:
    patch:
      operationId: changeIncidentStatus
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/IncidentStatusInput" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Incident" } } } }, "403": { description: Forbidden } }
...
  /contracts/{id}/payment:
    patch:
      operationId: approveContractPayment
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ContractPaymentInput" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { type: object, additionalProperties: true } } } }, "403": { description: Forbidden } }
  /company:
    patch:
      operationId: updateCompanyName
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CompanyUpdate" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Company" } } } }, "403": { description: Forbidden } }
  /members/{id}/role:
    patch:
      operationId: updateMemberRole
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MemberRoleInput" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { type: object, additionalProperties: true } } } }, "403": { description: Forbidden } }
...
  /settings/modules:
    patch:
      operationId: toggleModule
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ModuleUpdate" } } } }
      responses: { "200": { description: Updated, content: { application/json: { schema: { type: object, additionalProperties: true } } } }, "403": { description: Forbidden } }
  /cost-alerts/{id}/address:
    post:
      operationId: markCostAlertAddressed
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "204": { description: Addressed }, "403": { description: Forbidden } }
...
  /works/{id}/review:
    post:
      operationId: incrementWorkReviewed
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "204": { description: Counted }, "403": { description: Forbidden } }
components:
  schemas:
    CompanyOnboardingInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
...
    Company:
      type: object
      required: [id, name]
      properties: { id: { type: string }, name: { type: string } }
...
    Work:
      type: object
      required: [id, name, client, location, status, progress, budget, startDate, endDate, manager]
      properties:
        id: { type: string }
        name: { type: string }
        client: { type: string }
        location: { type: string }
        status: { type: string, enum: [planificación, en curso, pausada, completada] }
        progress: { type: integer }
        budget: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        manager: { type: string }
    WorkInput:
      type: object
      required: [name, client, location, status, progress, budget, startDate, endDate, manager]
      properties:
        name: { type: string }
        client: { type: string }
        location: { type: string }
        status: { type: string }
        progress: { type: integer }
        budget: { type: integer }
        startDate: { type: string }
        endDate: { type: string }
        manager: { type: string }
    Expense:
      type: object
      required: [id, workId, itemId, description, amount, type, date, vendor]
      properties:
        id: { type: string }
        workId: { type: string }
        itemId: { type: string }
        description: { type: string }
        amount: { type: integer, minimum: 1 }
        type: { type: string }
        date: { type: string }
        vendor: { type: string }
    ExpenseInput:
      type: object
      required: [workId, itemId, description, amount, type, date, vendor]
      properties:
        workId: { type: string }
        itemId: { type: string }
        description: { type: string }
        amount: { type: integer, minimum: 1 }
        type: { type: string }
        date: { type: string }
        vendor: { type: string }
    Incident:
      type: object
      required: [id, workId, title, description, status, priority, assignee, createdAt]
      properties:
        id: { type: string }
        workId: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string }
        priority: { type: string }
        assignee: { type: string }
        createdAt: { type: string }
...
    IncidentInput:
      type: object
      required: [workId, title, description, priority, assignee]
      properties:
        workId: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: string }
        assignee: { type: string }
    IncidentStatusInput:
      type: object
      required: [status]
      properties: { status: { type: string, enum: [abierto, en proceso, resuelto] } }
    ContractPaymentInput:
      type: object
      required: [amount]
      properties: { amount: { type: integer, minimum: 1 } }
    CompanyUpdate:
      type: object
      required: [name]
      properties: { name: { type: string, minLength: 1 } }
    MemberRoleInput:
      type: object
      required: [role]
      properties: { role: { type: string, enum: [owner_manager, office, site_manager] } }
...
    ModuleUpdate:
      type: object
      required: [module, enabled]
      properties: { module: { type: string, enum: [profitability, communications, subcontractors] }, enabled: { type: boolean } }
...